Rename A Necessary Risk section to fix spelling
