--- leblanc/components/home/ANecesaryRisk/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ANecesaryRisk'
