+++ leblanc/components/home/ANecessaryRisk/ANecessaryRisk.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useState } from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
-import s from './ANecesaryRisk.module.css'
+import s from './ANecessaryRisk.module.css'
 import { motion } from 'framer-motion'
 // import { LogoGusPena, LogoHorizontal, LogoANecesaryRisk } from '@leblanc/svg'
 import KidImg from 'public/assets/home/kid.png'
@@ -12,35 +12,34 @@ const ANecesaryRisk: FC = () => {
   const [toggle, setToggle] = useState(false)
   return (
     <div className={s.root}>
-    
-        <Image
-          priority
-          unoptimized
-          src={KidImg}
-          alt="kids leblanc"
-          className={s.bgImage}
-          width={1920}
-          height={1413}
-          layout="responsive"
-          objectFit="cover"
-        />
+      <Image
+        priority
+        unoptimized
+        src={KidImg}
+        alt="kids leblanc"
+        className={s.bgImage}
+        width={1920}
+        height={1413}
+        layout="responsive"
+        objectFit="cover"
+      />
 
       <motion.div
         animate={{ opacity: toggle ? 1 : 0 }}
         transition={{ duration: 0.35 }}
         className={s.bgImageWrapper}>
-          <a className={s.card}>
-            <Image
-              unoptimized
-              src={Kid2Img}
-              alt="kids leblanc"
-              className={s.cardImage}
-              width={1920}
-              height={1413}
-              layout="fill"
-              objectFit="cover"
-            />
-          </a>
+        <a className={s.card}>
+          <Image
+            unoptimized
+            src={Kid2Img}
+            alt="kids leblanc"
+            className={s.cardImage}
+            width={1920}
+            height={1413}
+            layout="fill"
+            objectFit="cover"
+          />
+        </a>
       </motion.div>
       <SwitchButton toggle={toggle} setToggle={setToggle} />
     </div>
