+++ leblanc/components/home/ANecessaryRisk/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ANecessaryRisk'
