+++ leblanc/components/home/index.ts
@@ -2,7 +2,7 @@ export { default as Hero } from './Hero'
 export { default as MenWomen } from './MenWomen'
 export { default as FeaturedCollection } from './FeaturedCollection'
 export { default as ArchivesSlider } from './ArchivesSlider'
-export { default as ANecesaryRisk } from './ANecesaryRisk'
+export { default as ANecessaryRisk } from './ANecessaryRisk'
 export { default as Navbar } from './Navbar'
 export { default as NavPanel } from './Navbar/NavPanel'
 export { default as NavDropdown } from './Navbar/NavDropdown'
